session5/pyproject.toml
[project]
name = "hotel-flight-supervisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph-supervisor>=0.0.31",
    "python-dotenv>=1.2.1",
]
